+++ CHANGELOG.md
@@ -2,6 +2,18 @@
 All major and minor version changes will be documented in this file. Details of
 patch-level version changes can be found in [commit messages](../../commits/master).
 
+## b_02.19_u1 - 19/02/2021
+- Bugfix endless set high-score
+- Visual indicator for locked level
+- high-score to endless
+- Credits screen
+- Made progress with achievements screen
+
+TO-DO
+- Complete achievements screen
+- Add total time played + number of objects to debug screen
+- Dev option hotkeys
+
 ## b_02.19 - 19/02/2021
 - Use `LevelCompleted` flag instead of `HighScores` to lock levels
 - Add high-scores to level select